+++ src/main/java/com/practiceproject/itmopracticeproject/boards/api/AdminBoardController.java
@@ -2,6 +2,7 @@ package com.practiceproject.itmopracticeproject.boards.api;
 
 import com.practiceproject.itmopracticeproject.boards.dto.BoardRequestDto;
 import com.practiceproject.itmopracticeproject.security.CurrentUser;
+import com.practiceproject.itmopracticeproject.user.db.GlobalRole;
 import com.practiceproject.itmopracticeproject.user.db.UserEntity;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
@@ -9,40 +10,49 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/boards")
-public class BoardController {
+@RequestMapping("/api/admin/boards")
+public class AdminBoardController {
 
     private final BoardService service;
 
-    public BoardController(BoardService service) {
+    public AdminBoardController(BoardService service) {
         this.service = service;
     }
 
     @PostMapping("/create")
     public ResponseEntity<?> createBoard(
             @Valid @RequestBody BoardRequestDto request,
-            @CurrentUser UserEntity user
-    ) {
+            @CurrentUser UserEntity user) {
+        if (user.getRole() != GlobalRole.ADMIN) {
+            throw new SecurityException("Access denied");
+        }
         return ResponseEntity
                 .status(HttpStatus.CREATED)
                 .body(service.createBoard(request, user.getId()));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getBoard(
-            @PathVariable("id") Long id,
-            @CurrentUser UserEntity user
-    ) {
-        return ResponseEntity.status(HttpStatus.OK).body(service.getBoardById(id, user));
+    public ResponseEntity<?> getBoardById(
+            @PathVariable("id") Long boardId,
+            @CurrentUser UserEntity user) {
+        if (user.getRole() != GlobalRole.ADMIN) {
+            throw new SecurityException("Access denied");
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).body(service.getBoardById(boardId, user));
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<?> updateBoard(
             @Valid @RequestBody BoardRequestDto request,
-            @PathVariable("id") Long boardId,
+            @PathVariable("id") Long id,
             @CurrentUser UserEntity user
     ) {
-        return ResponseEntity.status(HttpStatus.ACCEPTED).body(service.updateBoard(boardId, request, user));
+        if (user.getRole() != GlobalRole.ADMIN) {
+            throw new SecurityException("Access denied");
+        }
+
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body(service.updateBoard(id, request, user));
     }
 
     @DeleteMapping("/{id}")
@@ -50,6 +60,9 @@ public class BoardController {
             @PathVariable("id") Long id,
             @CurrentUser UserEntity user
     ) {
+        if (user.getRole() != GlobalRole.ADMIN) {
+            throw new SecurityException("Access denied");
+        }
         service.deleteBoardById(id, user);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
